requirements.txt
streamlit>=1.30
gspread>=6.0
google-auth>=2.0
pandas>=2.0
numpy>=1.26
openpyxl>=3.1
plotly>=5.0
